feat: generic function fetchApi
- L'implémentation de fetchpoke dépend maintenant de la fonction fetchApi.
- Ceci permet au passage de simplifier la manière dont fetchpoke est appelée, ce qui simplifie la lecture du code.
